flight-app-py/requirements.txt
Flask==3.0.3
flasgger==0.9.7.1

# OpenTelemetry packages
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-logging
opentelemetry-exporter-otlp
opentelemetry-instrumentation-wsgi
